# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用于描述漏洞扫描平台的后端 API。",
        "title": "VulnFusion API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/admin/results": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员查看所有任务的扫描结果",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "获取所有扫描结果",
                "responses": {
                    "200": {
                        "description": "所有扫描结果",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Result"
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/results/export/{task_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "导出某个任务的扫描结果（JSON 格式）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "导出任务结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务 ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "任务及其结果数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "导出失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/results/task/{task_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取某个任务 ID 下所有扫描结果（用户或管理员）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "获取任务扫描结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务 ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "扫描结果列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Result"
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "无权限访问",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定任务下所有扫描结果（用户本人或管理员）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "删除任务结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务 ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/results/{id}": {
            "get": {
                "description": "根据结果 ID 查询具体的漏洞信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Result"
                ],
                "summary": "查看单个扫描结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "扫描结果 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "扫描结果详情",
                        "schema": {
                            "$ref": "#/definitions/models.Result"
                        }
                    },
                    "400": {
                        "description": "参数格式错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "结果不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Result": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "详细信息（原始输出或解析后的内容）",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "severity": {
                    "description": "风险等级：low / medium / high / critical",
                    "type": "string"
                },
                "target": {
                    "description": "受影响目标",
                    "type": "string"
                },
                "taskID": {
                    "description": "所属任务 ID",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "记录时间",
                    "type": "string"
                },
                "vulnerability": {
                    "description": "漏洞名称或标识",
                    "type": "string"
                }
            }
        },
...
    }
}